# src/main/resources/openapi/open-api-swagger.yaml
# AI_INSTRUCTION:
# When modifying this file, use compact flow-style YAML to declare schemas and simple arrays.
# - Use inline arrays for required fields and tags: `required: [field1, field2]`
# - Use inline objects for simple properties: `propertyName: {type: string, format: int64, nullable: true}`
# - Use inline objects for nested single-item schemas: `schema: {type: array, items: {$ref: '...'}}`
# - Keep the file readable but minimize unnecessary vertical lines.

openapi: 3.0.3
info:
  title: Notification API Server
  version: 1.0.0
  description: Contract-First API specification for managing notifications

paths:
  /notifications:
    post:
      tags: [Notifications]
      summary: Create notification
      operationId: createNotification
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateNotificationRequest'}
      responses:
        '201':
          description: Notification created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NotificationResponse'}
        '400':
          description: Invalid request
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProblemDetail'}
        '500':
          description: Internal server error
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProblemDetail'}
  /notifications/stream:
    post:
      tags: [Notifications]
      summary: Stream notifications via SSE
      operationId: streamNotifications
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NotificationStreamRequest'}
      responses:
        '200':
          description: Notification event stream
          content:
            text/event-stream:
              schema: {type: array, items: {$ref: '#/components/schemas/NotificationResponse'}}
        '400':
          description: Invalid request
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProblemDetail'}
        '500':
          description: Internal server error
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProblemDetail'}

components:
  schemas:
    CreateNotificationRequest:
      type: object
      required: [title, message, type, priority, audienceType, severity]
      properties:
        title: {type: string, example: "System Maintenance"}
        message: {type: string, example: "Planned maintenance starts at 10 PM UTC."}
        actionUrl: {type: string, example: "https://example.org/status", nullable: true}
        type: {type: string, example: "MAINTENANCE"}
        priority: {type: string, enum: [LOW, NORMAL, HIGH, CRITICAL], example: HIGH}
        audienceType: {type: string, enum: [USER, ROLE, GLOBAL], example: ROLE}
        severity: {type: string, enum: [INFO, FAILURE, SUCCESS], example: INFO}
        targets: {type: array, items: {type: string}, nullable: true, example: ["OPS", "SUPPORT"]}
        expiresAt: {type: string, format: date-time, nullable: true}
    NotificationStreamRequest:
      type: object
      required: [id]
      properties:
        id: {type: string, example: "user-123"}
        roles: {type: array, items: {type: string}, nullable: true, example: ["OPS", "ADMIN"]}
    NotificationResponse:
      type: object
      required: [id, title, message, type, priority, audienceType, severity, status, createdAt, updatedAt]
      properties:
        id: {type: string, example: "01JZMWJ1KQ3S5N4Y1PEFAVX3R0"}
        title: {type: string, example: "System Maintenance"}
        message: {type: string, example: "Planned maintenance starts at 10 PM UTC."}
        actionUrl: {type: string, example: "https://example.org/status", nullable: true}
        type: {type: string, example: "MAINTENANCE"}
        priority: {type: string, enum: [LOW, NORMAL, HIGH, CRITICAL], example: HIGH}
        audienceType: {type: string, enum: [USER, ROLE, GLOBAL], example: ROLE}
        severity: {type: string, enum: [INFO, FAILURE, SUCCESS], example: INFO}
        status: {type: string, enum: [ACTIVE], example: ACTIVE}
        targets: {type: array, items: {type: string}, nullable: true, example: ["OPS", "SUPPORT"]}
        expiresAt: {type: string, format: date-time, nullable: true}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    ProblemDetail:
      type: object
      description: RFC 7807 Problem Details for error responses
      required: [type, title, status, detail]
      properties:
        type: {type: string, format: uri, example: "https://api.example.com/errors/server-error"}
        title: {type: string, example: "Server Error"}
        status: {type: integer, format: int32, example: 500}
        detail: {type: string, example: "An unexpected error occurred while processing the request"}
        instance: {type: string, format: uri, example: "/api/resource/123"}
        path: {type: string, example: "/api/resource"}